Fix logging in electron certa tests.
Essentially just reacting to this change in mocha: https://github.com/mochajs/mocha/pull/3725
