+++ common/changes/@bentley/certa/fix-certa-logging_2021-05-10-15-46.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/certa",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/certa",
+  "email": "33036725+wgoehrig@users.noreply.github.com"
+}
