+++ tools/certa/src/utils/initMocha.ts
@@ -6,23 +6,6 @@
 
 type CertaConfig = import("../CertaConfig").CertaConfig;
 declare let _CERTA_CONFIG: CertaConfig; // eslint-disable-line @typescript-eslint/naming-convention
-declare let _CertaConsole: undefined | ((name: string, args: any[]) => void); // eslint-disable-line @typescript-eslint/naming-convention
-
-// Redirect all console output back to the main (backend) process, if necessary
-if (typeof _CertaConsole !== "undefined") {
-  function forwardConsole(name: keyof typeof console) {
-    const original = console[name];
-    console[name] = (...args: any[]) => {
-      _CertaConsole!(name, args);
-      // Also preserve the original behavior. This way, test progress is reported in both the backend _and_ frontend processes.
-      // This helps keep the output readable when debugging the frontend.
-      original.apply(console, args);
-    };
-  }
-  forwardConsole("log");
-  forwardConsole("error");
-  forwardConsole("dir");
-}
 
 ((config: CertaConfig) => {
   const mochaOpts = config.mochaOptions;
